update image controller

+++ app/Http/Controllers/Api/ImageController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Api;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\Image;
+use App\Models\Product;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class ImageController extends Controller
 {
@@ -19,11 +21,15 @@ class ImageController extends Controller
     {
         try {
             $img = Image::findOrFail($id);
+            $product = Product::with('images')->findOrFail($img->product_id);
+            if (count($product->images) <= 1) {
+                return response(trans('product.update.delete_last_file_fail'), 400);
+            }
             $img->delete();
 
             return response()->json($img);
         } catch (ModelNotFoundException $e) {
-            return response()->setStatusCode(400);
+            return response(trans('messages.delete_fail'), 400);
         }
     }
 }
